Since very long time FreeBSD removed some obsolete functions from their ABI.  In order to compile some node modules on very old FreeBSD version some macros have to be passed to the compiler. Unfortunately the test used to check which FreeBSD version is installed on the host is quite poor since it is based on a text string which corresponds to the LLVM version. With FreeBSD version greater or equal to FreeBSD 12, LLVM major version is greater or equal to 10. Then the string comparison between "4.0" and the current LLVM version "10.xx" gives wrong result and unsupported MACROS are applied during compilation whereas it should not be applied.
On "newer" version of node (at least since node12) this test has been
removed. However since the current version of DocumentServer does not
compiles with node12 or greater (due to the use of V8 code which is no
more available in the V8 version contained in node12 and greater) we
need to patch the common.gypi file which is cached during npm install
process.
Once node_gyp files are cached other calls to npm install use the cached
version. So patching the cache files will work during the compilation
process.

To summarize this part of the code is only useful because
documentserver version use VERY VERY VERY VERY old V8 version. Once
DocumentServer will be able to use newer V8 version this hack will be
removed.